RA-1232 moved interfaces into appropriate folder as per code review.
