+++ src/Esfa.Vacancy.Register.Application/Esfa.Vacancy.Register.Application.csproj
@@ -65,17 +65,17 @@
   <ItemGroup>
     <Compile Include="Exceptions\UnauthorisedException.cs" />
     <Compile Include="Exceptions\ResourceNotFoundException.cs" />
-    <Compile Include="Interfaces\IFrameworkCodeConverter.cs" />
-    <Compile Include="Interfaces\IStandardCodeConverter.cs" />
     <Compile Include="Interfaces\ITrainingDetailService.cs" />
     <Compile Include="Interfaces\IApprenticeshipSearchService.cs" />
-    <Compile Include="Interfaces\IVacancySearchParametersConverter.cs" />
+    <Compile Include="Queries\SearchApprenticeshipVacancies\IVacancySearchParametersConverter.cs" />
     <Compile Include="Queries\GetVacancy\GetVacancyValidator.cs" />
     <Compile Include="Queries\GetVacancy\GetVacancyQueryHandler.cs" />
     <Compile Include="Queries\GetVacancy\GetVacancyRequest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Queries\GetVacancy\GetVacancyResponse.cs" />
     <Compile Include="Queries\SearchApprenticeshipVacancies\FrameworkCodeConverter.cs" />
+    <Compile Include="Queries\SearchApprenticeshipVacancies\IFrameworkCodeConverter.cs" />
+    <Compile Include="Queries\SearchApprenticeshipVacancies\IStandardCodeConverter.cs" />
     <Compile Include="Queries\SearchApprenticeshipVacancies\SearchApprenticeshipVacanciesQueryHandler.cs" />
     <Compile Include="Queries\SearchApprenticeshipVacancies\SearchApprenticeshipVacanciesRequest.cs" />
     <Compile Include="Queries\SearchApprenticeshipVacancies\SearchApprenticeshipVacanciesResponse.cs" />
