+++ src/Esfa.Vacancy.Register.Application/Queries/SearchApprenticeshipVacancies/IStandardCodeConverter.cs
@@ -1,8 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using Esfa.Vacancy.Register.Application.Queries.SearchApprenticeshipVacancies;
 
-namespace Esfa.Vacancy.Register.Application.Interfaces
+namespace Esfa.Vacancy.Register.Application.Queries.SearchApprenticeshipVacancies
 {
     public interface IStandardCodeConverter
     {
