+++ src/Esfa.Vacancy.Register.Application/Queries/SearchApprenticeshipVacancies/IVacancySearchParametersConverter.cs
@@ -1,8 +1,7 @@
 ﻿using System.Threading.Tasks;
-using Esfa.Vacancy.Register.Application.Queries.SearchApprenticeshipVacancies;
 using Esfa.Vacancy.Register.Domain.Entities;
 
-namespace Esfa.Vacancy.Register.Application.Interfaces
+namespace Esfa.Vacancy.Register.Application.Queries.SearchApprenticeshipVacancies
 {
     public interface IVacancySearchParametersConverter
     {
